Add email api
