+++ .env.example
@@ -10,3 +10,5 @@ GOOGLE_CLIENT_SECRET='your_google_client_secret'
 GOOGLE_REDIRECT_URL=http://localhost:${SERVER_PORT}/api/auth/google/callback
 CLIENT_URL=http://localhost:5173
 IMG_HOST=http://localhost:${SERVER_PORT}
+SMTP_MAIL=your_smtp_mail
+SMTP_PASSWORD=your_smtp_password
