+++ package.json
@@ -23,6 +23,7 @@
     "express": "^4.18.2",
     "jsonwebtoken": "^9.0.2",
     "multer": "^1.4.5-lts.1",
+    "nodemailer": "^6.9.13",
     "passport": "^0.6.0",
     "passport-google-oauth20": "^2.0.0",
     "passport-jwt": "^4.0.1",
@@ -41,6 +42,7 @@
     "@types/jsonwebtoken": "^9.0.2",
     "@types/multer": "^1.4.11",
     "@types/node": "^20.1.4",
+    "@types/nodemailer": "^6.4.15",
     "@types/passport": "^1.0.12",
     "@types/passport-google-oauth20": "^2.0.14",
     "@types/passport-jwt": "^3.0.9",
