+++ src/app.ts
@@ -14,6 +14,7 @@ import { CLIENT_URL } from './configs/envConfig'
 import cookieParser from 'cookie-parser'
 import menteeRouter from './routes/mentee/mentee.route'
 import fs from 'fs'
+import emailRouter from './routes/emails/emails.route'
 
 const app = express()
 const staticFolder = 'uploads'
@@ -39,6 +40,7 @@ app.use('/api/admin', adminRouter)
 app.use('/api/mentors', mentorRouter)
 app.use('/api/mentees', menteeRouter)
 app.use('/api/categories', categoryRouter)
+app.use('/api/emails', emailRouter)
 
 if (!fs.existsSync(staticFolder)) {
   fs.mkdirSync(staticFolder, { recursive: true })
