+++ src/configs/envConfig.ts
@@ -14,3 +14,5 @@ export const GOOGLE_CLIENT_SECRET = process.env.GOOGLE_CLIENT_SECRET ?? ''
 export const GOOGLE_REDIRECT_URL = process.env.GOOGLE_REDIRECT_URL ?? ''
 export const CLIENT_URL = process.env.CLIENT_URL ?? ''
 export const IMG_HOST = process.env.IMG_HOST ?? ''
+export const SMTP_MAIL = process.env.SMTP_MAIL ?? ''
+export const SMTP_PASS = process.env.SMTP_PASS ?? ''
